# Improved perf of SingleThreadEventExecutor