--- a/examples/Echo.Client/App.config
+++ b/examples/Echo.Client/App.config
@@ -1,12 +1,20 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
     <startup> 
         <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
     </startup>
   <appSettings>
-    <add key="ssl" value="True"/>
-    <add key="host" value="127.0.0.1"/>
-    <add key="port" value="8007"/>
-    <add key="size" value="256"/>
+    <add key="ssl" value="True" />
+    <add key="host" value="127.0.0.1" />
+    <add key="port" value="8007" />
+    <add key="size" value="256" />
   </appSettings>
+  <runtime>
+    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
+      <dependentAssembly>
+        <assemblyIdentity name="Newtonsoft.Json" publicKeyToken="30ad4fe6b2a6aeed" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-6.0.0.0" newVersion="6.0.0.0" />
+      </dependentAssembly>
+    </assemblyBinding>
+  </runtime>
 </configuration>